Fix for all warnings clean up

+++ src/adaptation/NfcAdaptation.cpp
@@ -567,7 +567,7 @@ void IoctlCallback(::android::hardware::nfc::V1_0::NfcData outputData) {
   const char* func = "IoctlCallback";
   nfc_nci_ExtnOutputData_t* pOutData =
       (nfc_nci_ExtnOutputData_t*)&outputData[0];
-  ALOGD("%s Ioctl Type=%llu", func, pOutData->ioctlType);
+  ALOGD("%s Ioctl Type=%llu", func, (unsigned long long)pOutData->ioctlType);
   NfcAdaptation* pAdaptation = (NfcAdaptation*)pOutData->context;
   /*Output Data from stub->Proxy is copied back to output data
    * This data will be sent back to libnfc*/
@@ -602,7 +602,7 @@ int NfcAdaptation::HalIoctl(long arg, void* p_data) {
   data.setToExternal((uint8_t*)pInpOutData, sizeof(nfc_nci_IoctlInOutData_t));
   if(mHalNxpNfc != nullptr)
       mHalNxpNfc->ioctl(arg, data, IoctlCallback);
-  ALOGD("%s Ioctl Completed for Type=%llu", func, pInpOutData->out.ioctlType);
+  ALOGD("%s Ioctl Completed for Type=%llu", func, (unsigned long long)pInpOutData->out.ioctlType);
   return (pInpOutData->out.result);
 }
